[Issue #57] revert intVector type and revise docs (#58)

+++ pixels-core/include/vector/LongColumnVector.h
@@ -11,7 +11,7 @@
 class LongColumnVector: public ColumnVector {
 public:
     long * longVector;
-    int * intVector;
+	long * intVector;
     /**
     * Use this constructor by default. All column vectors
     * should normally be the default size.
